requirements.txt
pytest
selenium
allure-pytest
configparser
allure-python-commons